fix(lint): resolve all 254 clippy warnings across 42 Rust source files
- Update MSRV from 1.77.2 to 1.80.0 (eliminates 75 incompatible_msrv warnings)
- Remove 53 needless borrows for generic arguments
- Simplify 40 redundant closures to function references
- Derive Default for 6 enums instead of manual impl
- Remove 9 unnecessary u64/u8 casts
- Collapse 10 else-if blocks and identical branches
- Replace 3 manual strip_prefix patterns
- Fix wrong_self_convention for Copy types (3 methods)
- Replace manual char comparisons with array syntax
- Add allow attributes for intentional type_complexity and too_many_arguments
- Replace floor_char_boundary (requires Rust 1.91) with MSRV-compatible alternative
- Rename XAI enum variant to Xai (serde rename_all="lowercase" preserves wire format)
- Add ignore rules for windows and rusqlite crates in dependabot

+++ src-tauri/src/archive_browse.rs
@@ -204,7 +204,7 @@ pub async fn list_7z(archive_path: String, password: Option<String>) -> Result<V
         let is_dir = entry.is_directory();
         ArchiveEntry {
             name,
-            size: entry.size() as u64,
+            size: entry.size(),
             compressed_size: 0,
             is_dir,
             is_encrypted: entry.has_stream() && secret_password.is_some(),
@@ -388,7 +388,7 @@ pub async fn list_rar(archive_path: String) -> Result<Vec<ArchiveEntry>, String>
         let entry = entry_result.map_err(|e| format!("Failed to read entry: {}", e))?;
         entries.push(ArchiveEntry {
             name: entry.filename.to_string_lossy().to_string(),
-            size: entry.unpacked_size as u64,
+            size: entry.unpacked_size,
             compressed_size: 0, // RAR crate doesn't expose packed_size directly
             is_dir: entry.is_directory(),
             is_encrypted: entry.is_encrypted(),
